requirements.txt
openai
GitPython
pytest
python-dotenv
chromadb
tiktoken
tree_sitter
langchain
networkx